+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Resolved the conflict between start/unstarr and select/deselect all in compund widgets in Browse panel [#1827](https://github.com/rokwire/illinois-app/issues/1827).
+### Added
+- Wellness To-Do home/favorite widget [#1828](https://github.com/rokwire/illinois-app/issues/1828).
 
 ## [4.0.37] - 2022-06-24
 ### Changed